Fixed merge issues
