--- Editor/Resources/EditorWindow/GameLiftPlugin.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e2c810574e47434dada447ccfcddc894
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
