+++ Editor/Resources/EditorWindow/GameLiftPlugin.uss
@@ -1,25 +1,12 @@
 /* Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. */
 /* SPDX-License-Identifier: Apache-2.0 */
+
 @import './common.uss';
 
 :root {
     height: 100%;
     width: 100%;
-    font-size: 16px;
-    
-    --color-menu-background: #333333;
-    --color-menu-border: #4c4c4c;
-    --color-content-background: #3F3F3F;
-    --color-button: #585858;
-    --color-button-primary: #2C5D87;
-    --color-button-hover: #386082;
-    --color-font-content: #d2d2d2;
-    
-    --font-size-menu: 18px;
-    --font-size-header: 32px;
-    --font-size-description: 16px;
-    
-    --content-width: 920px;
+    font-size: var(--font-size-default);
 }
 
 .main {
@@ -42,18 +29,13 @@
     background-color: var(--color-background-primary);
 }
 
-.tab__button {
-    font-size: var(--font-size-large);
-    background-color: var(--color-content-background);
-}
-
 .main__content .tab__content{
      padding: 24px;
      min-width: var(--content-width);
  }
 
 .tab__button {
-    font-size: var(--font-size-menu);
+    font-size: var(--font-size-large);
     -unity-text-align: middle-left;
     padding: 0 12px 0 20px;
     margin: 0 0 12px 0;
@@ -73,42 +55,9 @@
 
 .tab__content {
     display: none;
-    color: var(--color-font-content);
+    color: var(--color-font-default);
 }
 
 .tab__content.tab__content--selected {
     display: flex;
-}
-
-.tab__content__header {
-    font-size: var(--font-size-header);
-}
-
-.tab__content__description {
-    font-size: var(--font-size-description);
-}
-
-.tab__content > * > * {
-    margin-bottom: 12px;
-}
-
-Label {
-    white-space: normal;
-}
-
-.button {
-    border-radius: 4px;
-    border-width: 2px;
-    border-color: #242424;
-    padding: 6px 24px 6px 24px;
-    margin: 0;
-    background-color: var(--color-button);
-}
-
-.button--primary {
-    background-color: var(--color-button-primary);
-}
-
-.button:hover {
-    background-color: var(--color-button-hover);
 }
