+++ Editor/Resources/EditorWindow/Pages/LandingPage.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using AmazonGameLift.Editor;
+using Editor.Window;
 using UnityEngine;
 using UnityEngine.UIElements;
