+++ Editor/Resources/EditorWindow/Pages/LandingPage.uxml
@@ -1,7 +1,7 @@
 ﻿<ui:UXML xmlns:ui="UnityEngine.UIElements" >
     <Style src="./LandingPage.uss" />
-    <ui:Label name="LandingPageHeader" class="TabContent__header" text="Amazon GameLift"/>
-    <ui:Label name="LandingPageDescription" class="TabContent__description" text="Amazon GameLift is a dedicated game server hosting solution that deploys, operates, and scales cloud servers for multiplayer games.&#10;Whether you’re looking for a fully managed solution, or just the feature you need, GameLift leverages the power of AWS to deliver the best latency possible, low player wait times, and maximum cost savings."/>
+    <ui:Label name="LandingPageHeader" class="content__header" text="Amazon GameLift"/>
+    <ui:Label name="LandingPageDescription" class="content__description" text="Amazon GameLift is a dedicated game server hosting solution that deploys, operates, and scales cloud servers for multiplayer games.&#10;Whether you’re looking for a fully managed solution, or just the feature you need, GameLift leverages the power of AWS to deliver the best latency possible, low player wait times, and maximum cost savings."/>
     <ui:VisualElement name="Divider"  />        
     <ui:VisualElement class="cards">
         <ui:VisualElement class="card">
