+++ Editor/Resources/EditorWindow/common.uss
@@ -5,7 +5,47 @@
     --color-background-accent: #333333;
     --color-background-primary: #3F3F3F;
     --color-button-primary: #2C5D87;
+    --color-button-default: #585858;
     --color-button-hover: #386082;
-
+    --color-menu-border: #4c4c4c;
+    --color-font-default: #d2d2d2;
+    
+    --font-size-default: 16px;
     --font-size-large: 18px;
+    --font-size-massive: 32px;
+
+    --content-width: 920px;
+}
+
+.content__header {
+    font-size: var(--font-size-massive);
+}
+
+.content__description {
+    font-size: var(--font-size-default);
+}
+
+.tab__content > * > * {
+    margin-bottom: 12px;
+}
+
+Label {
+    white-space: normal;
+}
+
+.button {
+    border-radius: 4px;
+    border-width: 2px;
+    border-color: #242424;
+    padding: 6px 24px 6px 24px;
+    margin: 0;
+    background-color: var(--color-button-default);
+}
+
+.button--primary {
+    background-color: var(--color-button-primary);
+}
+
+.button:hover {
+    background-color: var(--color-button-hover);
 }
